+++ brood/executor.py
@@ -9,15 +9,16 @@ from rich.console import Console
 from brood.command import Event
 from brood.config import BroodConfig
 from brood.fanout import Fanout
-from brood.message import InternalMessage, Message
+from brood.message import InternalMessage, Message, Verbosity
 from brood.monitor import KillOthers, Monitor
 from brood.renderer import RENDERERS
 
 
 class Executor:
-    def __init__(self, config: BroodConfig, console: Console):
+    def __init__(self, config: BroodConfig, console: Console, verbosity: Verbosity):
         self.config = config
         self.console = console
+        self.verbosity = verbosity
 
         self.events: Fanout[Event] = Fanout()
         self.messages: Fanout[Message] = Fanout()
@@ -25,6 +26,7 @@ class Executor:
         self.renderer = RENDERERS[config.renderer.type](
             config=self.config.renderer,
             console=self.console,
+            verbosity=self.verbosity,
             events=self.events.consumer(),
             messages=self.messages.consumer(),
         )
@@ -51,9 +53,9 @@ class Executor:
     async def run(self) -> None:
         done, pending = await wait(
             (
-                self.monitor.run(),
-                self.renderer.mount(),
-                self.renderer.run(),
+                create_task(self.monitor.run(), name=f"Run {type(self.monitor).__name__}"),
+                create_task(self.renderer.mount(), name=f"Mount {type(self.renderer).__name__}"),
+                create_task(self.renderer.run(), name=f"Run {type(self.renderer).__name__}"),
             ),
             return_when=FIRST_EXCEPTION,
         )
@@ -72,27 +74,42 @@ class Executor:
     ) -> Optional[bool]:
         if exc_type:
             if exc_type is CancelledError:
-                text = f"Shutting down due to: keyboard interrupt"
+                await self.messages.put(
+                    InternalMessage(
+                        f"Shutting down due to: keyboard interrupt", verbosity=Verbosity.INFO
+                    )
+                )
             elif exc_type is KillOthers:
-                text = f"Shutting down due to: command failing"
+                await self.messages.put(
+                    InternalMessage(
+                        f"Shutting down due to: command failing", verbosity=Verbosity.INFO
+                    )
+                )
             else:
-                text = f"Shutting down due to: {exc_type.__name__}"
-            await self.messages.put(InternalMessage(text))
+                await self.messages.put(
+                    InternalMessage(
+                        f"Shutting down due to: {exc_type.__name__}: {exc_val}",
+                        verbosity=Verbosity.ERROR,
+                    )
+                )
 
         # Stop the monitor while repeatedly draining the renderer,
         # so that we can emit output during shutdown.
-        stop_monitor = create_task(self.monitor.stop())
-        drain_renderer = create_task(self.renderer.run(drain=True))
+        stop_monitor = create_task(self.monitor.stop(), name=f"Stop {type(self.monitor).__name__}")
+        drain_renderer = create_task(
+            self.renderer.run(drain=True), name=f"Drain {type(self.renderer)}"
+        )
         while True:
             done, pending = await wait((stop_monitor, drain_renderer), return_when=FIRST_COMPLETED)
-
             if stop_monitor in done:
                 await drain_renderer
                 break
             else:
                 await sleep(0.001)
-                drain_renderer = create_task(self.renderer.run(drain=True))
+                drain_renderer = create_task(
+                    self.renderer.run(drain=True), name=f"Drain {type(self.renderer).__name__}"
+                )
 
-        await self.renderer.unmount()
+        await create_task(self.renderer.unmount(), name=f"Unmount {type(self.renderer).__name__}")
 
         return True