+++ brood/renderer.py
@@ -3,10 +3,21 @@ from __future__ import annotations
 import asyncio
 import re
 import shutil
-from asyncio import ALL_COMPLETED, FIRST_EXCEPTION, Queue, create_task, wait
+from asyncio import (
+    ALL_COMPLETED,
+    FIRST_EXCEPTION,
+    AbstractEventLoop,
+    Queue,
+    all_tasks,
+    create_task,
+    current_task,
+    get_running_loop,
+    wait,
+)
 from dataclasses import dataclass, field
 from datetime import timedelta
 from functools import cached_property, partial
+from pathlib import Path
 from random import choice
 from shutil import get_terminal_size
 from typing import Dict, List, Literal, Mapping, Type
@@ -23,8 +34,10 @@ from rich.text import Text
 
 from brood.command import Command, Event, EventType
 from brood.config import CommandConfig, LogRendererConfig, RendererConfig
-from brood.message import CommandMessage, InternalMessage, Message
+from brood.message import CommandMessage, InternalMessage, Message, Verbosity
 from brood.utils import delay
+
+DIM_RULE = Rule(style="dim")
 
 NULL_STYLE = Style.null()
 RE_ANSI_ESCAPE = re.compile(r"(\x1b(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~]))")
@@ -71,6 +84,8 @@ class Renderer:
     config: RendererConfig
     console: Console
 
+    verbosity: Verbosity
+
     messages: Queue[Message]
     events: Queue[Event]
 
@@ -86,8 +101,12 @@ class Renderer:
     async def run(self, drain: bool = False) -> None:
         done, pending = await wait(
             (
-                create_task(self.handle_events(drain=drain)),
-                create_task(self.handle_messages(drain=drain)),
+                create_task(
+                    self.handle_events(drain=drain), name=f"{type(self).__name__} event handler"
+                ),
+                create_task(
+                    self.handle_messages(drain=drain), name=f"{type(self).__name__} message handler"
+                ),
             ),
             return_when=ALL_COMPLETED if drain else FIRST_EXCEPTION,
         )
@@ -123,7 +142,8 @@ class Renderer:
             message = await self.messages.get()
 
             if isinstance(message, InternalMessage):
-                await self.handle_internal_message(message)
+                if message.verbosity <= self.verbosity:
+                    await self.handle_internal_message(message)
             elif isinstance(message, CommandMessage):
                 await self.handle_command_message(message)
 
@@ -178,18 +198,16 @@ class LogRenderer(Renderer):
     def live(self) -> Live:
         return Live(
             console=self.console,
+            renderable=StatusTable(
+                loop=get_running_loop(),
+                status_bars=self.status_bars,
+                show_task_status=self.verbosity.is_debug,
+            ),
             auto_refresh=True,
             refresh_per_second=10,
             transient=True,
             screen=False,
         )
-
-    def update_live(self) -> None:
-        table = Table.grid(expand=True)
-        for k, v in sorted(self.status_bars.items(), key=lambda kv: kv[0].process.pid):
-            table.add_row(v)  # type: ignore
-
-        self.live.update(Group(Rule(style="dim"), table), refresh=True)
 
     async def mount(self) -> None:
         if not self.config.status_tracker:
@@ -224,8 +242,6 @@ class LogRenderer(Renderer):
 
         self.status_bars[event.manager] = p
 
-        self.update_live()
-
     async def handle_stopped_event(self, event: Event) -> None:
         if not self.config.status_tracker:
             return
@@ -241,12 +257,16 @@ class LogRenderer(Renderer):
         )
         p.update(TaskID(0), completed=1)
 
-        self.stop_tasks.append(delay(10, partial(self.remove_status_bar, manager=event.manager)))
+        self.stop_tasks.append(
+            delay(
+                delay=10,
+                fn=partial(self.remove_status_bar, manager=event.manager),
+                name=f"Remove status entry for pid {event.manager.process.pid}",
+            )
+        )
 
     async def remove_status_bar(self, manager: Command, delay: int = 10) -> None:
         self.status_bars.pop(manager, None)
-
-        self.update_live()
 
     async def handle_internal_message(self, message: InternalMessage) -> None:
         self.console.print(self.render_internal_message(message), soft_wrap=True)
@@ -285,6 +305,40 @@ class LogRenderer(Renderer):
             ),
             style=message.command_config.prefix_style or self.config.prefix_style,
         )
+
+
+@dataclass(frozen=True)
+class StatusTable:
+    loop: AbstractEventLoop
+    status_bars: Dict[Command, Progress]
+    show_task_status: bool
+
+    def __rich__(self) -> ConsoleRenderable:
+        table = Table.grid(expand=False, padding=(0, 1))
+        for k, v in sorted(self.status_bars.items(), key=lambda kv: kv[0].process.pid):
+            table.add_row(v)  # type: ignore
+
+        tables = [table]
+
+        if self.show_task_status:
+            debug_table = Table.grid(expand=False, padding=(0, 1))
+            active_task = current_task(self.loop)
+            for task in sorted(all_tasks(self.loop), key=lambda task: task.get_name()):
+                frame = task.get_stack(limit=1)[0]
+                code = frame.f_code
+
+                debug_table.add_row(
+                    Text(task.get_name()),
+                    Text(f"{Path(code.co_filename).name}:{frame.f_lineno}::{code.co_name}"),
+                    style=Style(dim=task is not active_task),
+                )
+
+            tables.append(debug_table)
+
+        ubertable = Table.grid(expand=True, padding=(0, 2))
+        ubertable.add_row(*tables)
+
+        return Group(DIM_RULE, ubertable)
 
 
 RENDERERS: Mapping[Literal["null", "log"], Type[Renderer]] = {